authCRUD/requirements.txt
djangorestframework-simplejwt==3.3
django mysqlclient
djangorestframework==3.10.3
django-cors-headers==3.1.1
